requirements.txt
astroid==2.5.6
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cryptography==3.3.2
google==3.0.0
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleAPI==0.1.0
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mailchimp3==3.0.14
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.8.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
stripe==2.58.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
pyjwt==2.4.0
